+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Verbatim filter for pandoc.
+
 ## [0.2.0] - 2025-03-19
 
 ### Added